Refactor chat navigation and add chat details screen
