+++ src/navigation/Chat/Chat.tsx
@@ -1,0 +1,5 @@
+import ChatListManagement from './ChatListManagement/ChatListManagement'
+
+export default function Chat(): JSX.Element {
+  return <ChatListManagement />
+}
