+++ src/navigation/Chat/ChatConversationManagement/ChatConversationManagement.tsx
@@ -1,13 +1,13 @@
 import React, { useState, useEffect, useCallback } from 'react'
 import { GiftedChat } from 'react-native-gifted-chat'
-import { useAppSelector } from '../store/store'
-import { selectedUser, selectedUserToken } from '../features/userSlice'
-import { ROUTE_API } from '../constants/api'
+import { useAppSelector } from '../../../store/store'
+import { selectedUser, selectedUserToken } from '../../../features/userSlice'
+import { ROUTE_API } from '../../../constants/api'
 import axios from 'axios'
 import { IMessage } from 'react-native-gifted-chat'
-import { socket } from '../socket'
+import { socket } from '../../../socket'
 
-export default function Chat() {
+export default function ChatConversationManagement() {
   const userId = useAppSelector(selectedUser).user_id
   const token = useAppSelector(selectedUserToken)
   const [messages, setMessages] = useState<IMessage[]>([])
