+++ src/navigation/Chat/ChatListManagement/ChatListManagement.tsx
@@ -1,0 +1,5 @@
+import { Text } from 'react-native'
+
+export default function ChatListManagement(): JSX.Element {
+  return <Text>ChatListManagement</Text>
+}
